Reactify Large Capital Profile location section
