--- src/apps/companies/apps/investments/large-capital-profile/client/state.js
@@ -1,2 +1,0 @@
-export const TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION =
-  'TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION'
