--- src/apps/companies/apps/investments/large-capital-profile/client/tasks.js
@@ -1,9 +1,0 @@
-import { apiProxyAxios } from '../../../../../../client/components/Task/utils'
-
-export const updateInvestorLocation = (values) => {
-  const { id, investor_company_id } = values
-  return apiProxyAxios.patch(
-    `v4/large-investor-profile/${id}?investor_company_id=${investor_company_id}`,
-    values
-  )
-}
