+++ src/client/modules/Companies/CompanyInvestments/LargeCapitalProfile/state.js
@@ -1,2 +1,4 @@
 export const TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS =
   'TASK_SAVE_LARGE_CAPITAL_INVESTOR_DETAILS'
+export const TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION =
+  'TASK_SAVE_LARGE_CAPITAL_INVESTOR_LOCATION'
