+++ src/client/modules/Companies/CompanyInvestments/LargeCapitalProfile/transformers.js
@@ -6,6 +6,13 @@ import {
 const checkCurrencyField = (currencyField) =>
   currencyField || currencyField <= 0 ? currencyField : null
 
+const getName = (obj) => obj.name
+
+export const transformArray = (arr = []) => arr.map(getName).join(', ')
+
+export const transformValueToId = (values = []) =>
+  values.map((value) => ({ id: value.value }))
+
 export const transformInvestorDetailsToApi = ({
   profileId,
   companyId,
@@ -35,5 +42,27 @@ export const transformInvestorDetailsToApi = ({
     required_checks_conducted_by: adviser
       ? transformOption(adviser)
       : undefined,
+  }
+}
+
+export const transformLocationDetailsToApi = ({
+  profileId,
+  companyId,
+  values,
+}) => {
+  const {
+    uk_region_locations,
+    other_countries_being_considered,
+    notes_on_locations,
+  } = values
+
+  return {
+    id: profileId,
+    investor_company_id: companyId,
+    uk_region_locations: transformValueToId(uk_region_locations),
+    other_countries_being_considered: transformValueToId(
+      other_countries_being_considered
+    ),
+    notes_on_locations,
   }
 }
